+++ lib/posix-socket/socket.c
@@ -93,7 +93,7 @@ static struct uk_ofile *socketfd_get(int fd)
 
 static ssize_t
 socket_read(const struct uk_file *sock,
-	    const struct iovec *iov, int iovcnt,
+	    const struct iovec *iov, size_t iovcnt,
 	    size_t off, long flags __unused)
 {
 	ssize_t ret;
@@ -125,7 +125,7 @@ socket_read(const struct uk_file *sock,
 
 static ssize_t
 socket_write(const struct uk_file *sock,
-	     const struct iovec *iov, int iovcnt,
+	     const struct iovec *iov, size_t iovcnt,
 	     size_t off, long flags __unused)
 {
 	ssize_t ret;
