+++ lib/posix-tty/pseudo.c
@@ -22,7 +22,8 @@ static const char VOID_VOLID[] = "void_vol";
 static const char ZERO_VOLID[] = "zero_vol";
 
 static ssize_t null_read(const struct uk_file *f __maybe_unused,
-			 const struct iovec *iov __unused, int iovcnt __unused,
+			 const struct iovec *iov __unused,
+			 size_t iovcnt __unused,
 			 size_t off __unused, long flags __unused)
 {
 	UK_ASSERT(f->vol == NULL_VOLID);
@@ -30,7 +31,8 @@ static ssize_t null_read(const struct uk_file *f __maybe_unused,
 }
 
 static ssize_t void_read(const struct uk_file *f __maybe_unused,
-			 const struct iovec *iov __unused, int iovcnt __unused,
+			 const struct iovec *iov __unused,
+			 size_t iovcnt __unused,
 			 size_t off __unused, long flags __unused)
 {
 	UK_ASSERT(f->vol == VOID_VOLID);
@@ -38,14 +40,14 @@ static ssize_t void_read(const struct uk_file *f __maybe_unused,
 }
 
 static ssize_t zero_read(const struct uk_file *f __maybe_unused,
-			 const struct iovec *iov, int iovcnt,
+			 const struct iovec *iov, size_t iovcnt,
 			 size_t off __unused, long flags __unused)
 {
 	ssize_t total = 0;
 
 	UK_ASSERT(f->vol == ZERO_VOLID);
 
-	for (int i = 0; i < iovcnt; i++) {
+	for (size_t i = 0; i < iovcnt; i++) {
 		if (unlikely(!iov[i].iov_base && iov[i].iov_len))
 			return -EFAULT;
 		memset(iov[i].iov_base, 0, iov[i].iov_len);
@@ -55,7 +57,7 @@ static ssize_t zero_read(const struct uk_file *f __maybe_unused,
 }
 
 static ssize_t null_write(const struct uk_file *f __maybe_unused,
-			  const struct iovec *iov, int iovcnt,
+			  const struct iovec *iov, size_t iovcnt,
 			  size_t off __unused, long flags __unused)
 {
 	ssize_t total = 0;
@@ -63,7 +65,7 @@ static ssize_t null_write(const struct uk_file *f __maybe_unused,
 	UK_ASSERT(f->vol == NULL_VOLID || f->vol == ZERO_VOLID ||
 		  f->vol == VOID_VOLID);
 
-	for (int i = 0; i < iovcnt; i++)
+	for (size_t i = 0; i < iovcnt; i++)
 		total += iov[i].iov_len;
 	return total;
 }
