+++ lib/posix-tty/serial.c
@@ -88,7 +88,7 @@ static inline __ssz _console_out(const char *buf, __sz len)
 }
 
 static ssize_t serial_read(const struct uk_file *f,
-			   const struct iovec *iov, int iovcnt,
+			   const struct iovec *iov, size_t iovcnt,
 			   size_t off, long flags __unused)
 {
 	ssize_t total = 0;
@@ -100,7 +100,7 @@ static ssize_t serial_read(const struct uk_file *f,
 	if (!uk_file_poll_immediate(f, UKFD_POLLIN))
 		return 0;
 
-	for (int i = 0; i < iovcnt; i++) {
+	for (size_t i = 0; i < iovcnt; i++) {
 		char *buf = iov[i].iov_base;
 		size_t len = iov[i].iov_len;
 		char *last;
@@ -137,7 +137,7 @@ static ssize_t serial_read(const struct uk_file *f,
 }
 
 static ssize_t serial_write(const struct uk_file *f __maybe_unused,
-			    const struct iovec *iov, int iovcnt,
+			    const struct iovec *iov, size_t iovcnt,
 			    size_t off, long flags __unused)
 {
 	ssize_t total = 0;
@@ -146,7 +146,7 @@ static ssize_t serial_write(const struct uk_file *f __maybe_unused,
 	if (unlikely(off))
 		return -ESPIPE;
 
-	for (int i = 0; i < iovcnt; i++) {
+	for (size_t i = 0; i < iovcnt; i++) {
 		char *buf = iov[i].iov_base;
 		size_t len = iov[i].iov_len;
 		int bytes_written;
